Top text; bug fix for using groups
A bug relating to the use of Timeouts and Flag/Reviews
and Groups in OBS has been fixed.
These features rely on changing visibility, which OBS does in the
context of a scene. Groups act as a virtual scene, so you must supply
the group name as a scene/container name for those.

Signed-off-by: Blake <blake@1024256.xyz>